Move init var to utils and add logs subdirectory
