+++ pkg/local/install.go
@@ -3,48 +3,10 @@ package local
 import (
 	"fmt"
 	"os"
-	"os/exec"
-	"path"
 	"strings"
 
-	"github.com/sirupsen/logrus"
-	log "github.com/sirupsen/logrus"
 	"github.com/tensorleap/cli-go/pkg/helm"
 )
-
-func InitVarDir() error {
-	_, err := os.Stat(VAR_DIR)
-	if os.IsNotExist(err) {
-		log.Printf("Creating directory: %s (you may be asked to enter the root user password)", VAR_DIR)
-		mkdirCmd := exec.Command("/bin/sh", "-c", fmt.Sprintf("sudo mkdir -p %s", VAR_DIR))
-		if err := mkdirCmd.Run(); err != nil {
-			return err
-		}
-
-		log.Print("Setting directory permissions")
-		chmodCmd := exec.Command("/bin/sh", "-c", fmt.Sprintf("sudo chmod -R 777 %s", VAR_DIR))
-		if err := chmodCmd.Run(); err != nil {
-			return err
-		}
-	} else if err != nil {
-		return err
-	}
-
-	for _, dir := range []string{"storage", "registry"} {
-		fullPath := path.Join(VAR_DIR, dir)
-		_, err := os.Stat(fullPath)
-		if os.IsNotExist(err) {
-			log.Printf("Creating directory: %s", fullPath)
-			if err := os.MkdirAll(fullPath, 0777); err != nil {
-				return err
-			}
-		} else if err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
 
 func InitDataVolumeDir(dataVolume string) error {
 	dataPath := strings.Split(dataVolume, ":")[0]
@@ -56,8 +18,8 @@ func GetDefaultDataVolume() string {
 	return fmt.Sprintf("%s:%s", defaultDataPath, defaultDataPath)
 }
 
-func InstallHelm(useGpu bool, dataContainerPath string, logger *logrus.Logger) error {
-	helmConfig, err := helm.CreateHelmConfig(KUBE_CONTEXT, KUBE_NAMESPACE, logger)
+func InstallHelm(useGpu bool, dataContainerPath string) error {
+	helmConfig, err := helm.CreateHelmConfig(KUBE_CONTEXT, KUBE_NAMESPACE)
 	if err != nil {
 		return err
 	}
